+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## main / unreleased
+
+* [ENHANCEMENT] Add support for RefID #3
+
 ## 0.1.0 / 2022-03-20
 
 Initial release.